Allow `prorate: false` on Subscription.update
Hi there,

Stripe allows you to send `prorate: false` in the post body on Subscription update so that they don't automatically prorate when a customer changes plans. We did this in stripity_stripe 1.6, and I'd like to keep doing it in 2.0.

Have tested and it works for me. Would love to know what you think.

Thanks!