Add note to Testing section about skipping starting stripe-mock
This variable can be discovered by digging in to the source, or found in
the github workflow files, but it's useful enough that I think it should
be in the docs.

Also added a quick way to start stripe-mock with docker.  The auto-start
did not work for me so I went down a small rabbit hole trying to figure
out how to get them running.