Add 'balance_transaction' property missing from dispute object
This property is not mentioned in the documentation, so I'm wondering if
it's been deprecated in favor of the array version (called
balance_transactions).  Unfortunately I ran into a scenario where the
array was empty but there was an ID here :-(

I figured I'd send the PR to start a conversation.  Do we want to include
this property or leave it off?   I think my position is to leave it off
since it's not mentioned in the docs, but I could very easily be talked
into the other position.